requirements.txt
aiokafka==0.8.0
async-timeout==4.0.2
asyncio==3.4.3
boto3==1.26.45
botocore==1.29.45
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
docopt==0.6.2
Flask==2.2.2
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsonify==0.5
kafka-python==2.0.2
MarkupSafe==2.1.1
packaging==22.0
pipreqs==0.4.11
python-dateutil==2.8.2
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.13
uvicorn==0.20.0
Werkzeug==2.2.2
yarg==0.1.9
